chore: initial nodes updated
