+++ src/app/(app)/(dashboard)/dashboard/project/[...route]/page.tsx
@@ -1,9 +1,7 @@
 import configPromise from '@payload-config'
-import { Service } from '@payload-types'
 import { getPayloadHMR } from '@payloadcms/next/utilities'
 
 import ProjectDetailsView from '@/components/dashboard/project'
-import NodesTesting from '@/components/dashboard/project/NodesTesting'
 import withAuth from '@/utils/withAuth'
 
 const payload = await getPayloadHMR({ config: configPromise })
@@ -19,14 +17,7 @@ const page = async ({ params }: { params: Promise<{ route: any }> }) => {
     },
   })
 
-  console.log('server', services)
-  // return <ProjectDetailsView slug={slug} />
-
-  return (
-    <NodesTesting slug={slug} services={services?.docs as Service[]}>
-      <ProjectDetailsView slug={slug} />
-    </NodesTesting>
-  )
+  return <ProjectDetailsView slug={slug} services={services?.docs} />
 }
 
 export default withAuth(page)
